+++ tests/trainer/logging_/test_logger_connector.py
@@ -204,7 +204,7 @@ def test_epoch_results_cache_dp(tmpdir):
         default_root_dir=tmpdir, accelerator="dp", gpus=2, limit_train_batches=2, limit_val_batches=2, max_epochs=1
     )
     trainer.fit(model)
-    trainer.test(model, ckpt_path=None)
+    trainer.test(model)
 
 
 def test_can_return_tensor_with_more_than_one_element(tmpdir):